requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
pydantic>=2.5.0
tomli>=2.0.0
tomli-w>=1.0.0
psutil>=5.9.0
pikpakapi
httpx
feedparser
telethon>=1.41,<2.0
requests>=2.32,<3.0
qrcode>=8.2,<9.0
psycopg2-binary>=2.9,<3.0
